% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedPriceR.R
\name{fixed_convert_currencies}
\alias{fixed_convert_currencies}
\title{fixed_convert_currencies}
\usage{
fixed_convert_currencies(
  price_start,
  from,
  to,
  date = lubridate::today(),
  floor_unit = "day"
)
}
\arguments{
\item{price_start}{A vector of values in 'from' currency.}

\item{from}{A currency to convert from.}

\item{to}{A currency to convert to.}

\item{date}{A date for currency coversion.}

\item{floor_unit}{See priceR::convert_currencies}
}
\value{
A vector of values in 'to' currency
}
\description{
fixed_convert_currencies
}
